requirements.txt
twitch-python
Pillow
